requirements.txt
boto3
datetime
pymongo
celery 
python-dotenv
pandas